作业

+++ Week02/models/user.go
@@ -1,6 +1,6 @@
 package models
 
 type User struct {
-	Id   int `db:"id"`
+	Id   int64 `db:"id"`
 	Name string `db:"name"`
 }
